Support existing indices option, move remaining state
